requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
mpmath==1.3.0
Naked==0.1.32
numpy==1.26.2
pycryptodome==3.19.0
PyYAML==6.0.1
requests==2.31.0
shellescape==3.8.1
sympy==1.12
urllib3==2.1.0
